requirements.txt
fastapi
pymongo[srv]
xrpl-py
lupa